requirements.txt
# Core data processing and ML
pandas~=2.1.0
numpy~=1.24.0
scikit-learn~=1.3.0

# Machine Learning and Deep Learning
tensorflow~=2.13.0

# Configuration management
pyyaml~=6.0

# Property-based testing
hypothesis~=6.88.0

# Optional: TensorFlow.js export (for web deployment)
tensorflowjs~=4.11.0

# Development and testing (optional)
pytest~=7.4.0
pytest-cov~=4.1.0
flake8~=6.0.0
black~=23.7.0